+++ source/renderer/modules/markdown-editor/hooks/heading-classes.js
@@ -27,22 +27,10 @@ module.exports = (cm) => {
 }
 
 function applyHeadingClasses (cm) {
-  let wrapperClass = ''
-  let needsRefresh = false // Will be set to true if at least one line has been altered
-
   // We'll only render the viewport
   const viewport = cm.getViewport()
   for (let i = viewport.from; i < viewport.to; i++) {
-    let oldClass = ''
     const line = cm.getLine(i)
-
-    // Retrieve the wrapper class
-    wrapperClass = cm.lineInfo(i).wrapClass
-
-    // Save the old class name
-    if (/size-header-\d/.test(wrapperClass)) {
-      oldClass = /(size-header-\d)/.exec(wrapperClass)[1]
-    }
 
     // Then remove all header styles
     for (let x = 1; x < 7; x++) {
@@ -51,8 +39,6 @@ function applyHeadingClasses (cm) {
 
     // Only re-apply a header class if allowed.
     if (cm.getModeAt({ 'line': i, 'ch': 0 }).name !== 'markdown') {
-      // Indicate a refresh if necessary
-      if (oldClass !== '') needsRefresh = true
       continue
     }
 
@@ -60,9 +46,6 @@ function applyHeadingClasses (cm) {
     let match = /^(#{1,6}) /.exec(line)
     if (match) {
       cm.addLineClass(i, 'wrap', `size-header-${match[1].length}`)
-      // If the new header class is different
-      // than the old one, indicate a refresh.
-      if (oldClass !== `size-header-${match[1].length}`) needsRefresh = true
     }
 
     if (i === 0) continue // No need to check for Setext header
@@ -103,18 +86,6 @@ function applyHeadingClasses (cm) {
       for (let line = begin; line <= i; line++) {
         cm.addLineClass(line, 'wrap', `size-header-${level}`)
       }
-
-      // If the new header class is different
-      // than the old one, indicate a refresh.
-      if (oldClass !== `size-header-${level}`) {
-        needsRefresh = true
-      }
     }
-  }
-
-  // If at least one header class has been altered, refresh the codemirror
-  // instance as the sizes won't match up in that case.
-  if (needsRefresh) {
-  //   cm.refresh()
   }
 }